Move line diff tests
